requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
openai>=1.30.0
anthropic>=0.40.0
python-docx>=1.1.0
python-pptx>=0.6.23
pdfplumber>=0.11.0
python-dotenv>=1.0.0
